Split only by line of dashes
